block-volume: moved common code to a utils package
